fix(tests): compile harness on linux
